Template instantiation rampage (much lighter common headers)
